Move tmp key creation to service schema utils
